perf(sand): create noise upfront and sample it later
Previously noise was created every frame. The end result looked good,
but the simulation could not maintain an acceptable framerate. This
commit makes it so a large sample of noise is created upfront and sampled
later. The overall continuity of the noise is worse; however, the
performance of the simulation has increased significantly.