fix(sand): load wasm file idiomatically
The wasm file is now included in the final build as a result of using
`?url` at the end of its import.

By placing the initialization logic in a `<script context="module">`
block, the file is loaded when the module is first evaluated. Moreover,
the block's `load` function exposes the proper `fetch` function to use
when initializing the file. In order to use the proper function I added
a svelte specific initialization function to the existing sand module.

Finally, the file's URL must include the domain name when initializing
the file on the server. I thought the proper `fetch` function would
resolve this, but I could not get it to work with the full domain
name...